// backend/src/shared/enums/roles.enum.ts
export enum Role {
  FARMER = 'FARMER',
  CHECKIN_OPERATOR = 'CHECKIN_OPERATOR',
  WEIGHING_OPERATOR = 'WEIGHING_OPERATOR',
  QUALITY_OPERATOR = 'QUALITY_OPERATOR',
  PROCUREMENT_OPERATOR = 'PROCUREMENT_OPERATOR',
  CENTRE_ADMIN = 'CENTRE_ADMIN',
  DISTRICT_ADMIN = 'DISTRICT_ADMIN',
  STATE_ADMIN = 'STATE_ADMIN',
  GOVERNMENT_ADMIN = 'GOVERNMENT_ADMIN',
  AUDITOR = 'AUDITOR',
  SYSTEM_ADMIN = 'SYSTEM_ADMIN',
}